requirements.txt
Django==4.2.17
django-jazzmin
Pillow
django-ckeditor
